xrpl-airdrop-tool/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
xrpl-py==2.4.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
websockets>=10.0,<11.0
cryptography==41.0.7
